fix proposal link
